# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Сервис предоставляет функционал для работы с сегментами и пользователями для целей аналитики",
        "title": "Сервис по работе с сегментами",
        "contact": {
            "name": "Грищенко Владимир",
            "email": "vladimirxsky@gmail.com"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/api/v1/users": {
            "get": {
                "description": "Возвращает список абсолютно всех пользователей",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Получить список всех пользователей",
                "responses": {
                    "200": {
                        "description": "Список всех пользователей",
                        "schema": {
                            "$ref": "#/definitions/internal_controller_http_v1.GetAllUsersResponse"
                        }
                    },
                    "400": {
                        "description": "Ошибка валидации данных запроса",
                        "schema": {
                            "$ref": "#/definitions/avito-rest-api_internal_error.ErrUserValidationError"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/avito-rest-api_internal_error.ErrInternalServerError"
                        }
                    }
                }
            },
            "post": {
                "description": "Создаёт пользователя на основе информации в теле запроса",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Создать пользователя",
                "parameters": [
                    {
                        "description": "Структура с информацией о создаваемом пользователе",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/avito-rest-api_internal_service.UserCreateInput"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Идентификатор созданного пользователя",
                        "schema": {
                            "$ref": "#/definitions/internal_controller_http_v1.UserCreateResponse"
                        }
                    },
                    "400": {
                        "description": "Ошибка валидации данных запроса",
                        "schema": {
                            "$ref": "#/definitions/avito-rest-api_internal_error.ErrUserValidationError"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/avito-rest-api_internal_error.ErrInternalServerError"
                        }
                    }
                }
            }
        },
        "/api/v1/users/addUserToSegments": {
            "post": {
                "description": "Добавляет пользователя с указанным ID в указанные сегменты",
                "tags": [
                    "users"
                ],
                "summary": "Добавить пользователя в сегменты",
                "parameters": [
                    {
                        "description": "Структура, содержащая ID пользователя и наименование сегментов, в которые необходимо добавить пользователя. Поле `end_date` у сегмента является опциональным, и, если не  установлено, сигнализирует о том, что время выхода пользователя из сегмента не определено (пока сегмент  не будет удалён или пользователь не будет удалён из этого сегмента)",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_controller_http_v1.AddUserToSegmentsInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Сообщение об успехе",
                        "schema": {
                            "$ref": "#/definitions/internal_controller_http_v1.AddUserToSegmentsResponse"
                        }
                    },
                    "400": {
                        "description": "Ошибка валидации данных запроса",
                        "schema": {
                            "$ref": "#/definitions/avito-rest-api_internal_error.ErrUserValidationError"
                        }
                    },
                    "404": {
                        "description": "Пользователь с указанным ID не был найден или некоторые из указанных сегментов не существуют",
                        "schema": {
                            "$ref": "#/definitions/avito-rest-api_internal_error.ErrUserNotFound"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/avito-rest-api_internal_error.ErrInternalServerError"
                        }
                    }
                }
            }
        },
        "/api/v1/users/deleteUserFromSegments": {
            "post": {
                "description": "Удаляет пользователя с указанным ID из указанных сегментов",
                "tags": [
                    "users"
                ],
                "summary": "Удалить пользователя из сегментов",
                "parameters": [
                    {
                        "description": "Структура, содержащая ID пользователя и наименования сегментов, из которых пользователя необходимо удалить",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_controller_http_v1.DeleteUserFromSegmentsInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Сообщение об успехе",
                        "schema": {
                            "$ref": "#/definitions/internal_controller_http_v1.DeleteUserFromSegmentsResponse"
                        }
                    },
                    "400": {
                        "description": "Ошибка валидации данных запроса, может возникать, если пользователь не входит в указанные сегменты",
                        "schema": {
                            "$ref": "#/definitions/avito-rest-api_internal_error.ErrUserValidationError"
                        }
                    },
                    "404": {
                        "description": "Пользователь с указанным ID не был найден или некоторые из указанных сегментов не существуют",
                        "schema": {
                            "$ref": "#/definitions/avito-rest-api_internal_error.ErrUserNotFound"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/avito-rest-api_internal_error.ErrInternalServerError"
                        }
                    }
                }
            }
        },
        "/api/v1/users/withSegments": {
            "get": {
                "description": "Возвращает список всех пользователей, включая список активных для каждого пользователя сегментов на момент совершения запроса",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Получить список всех пользователей, включая их сегменты",
                "responses": {
                    "200": {
                        "description": "Список пользователей с их активными сегментами",
                        "schema": {
                            "$ref": "#/definitions/internal_controller_http_v1.GetAllUsersWithSegmentsResponse"
                        }
                    },
                    "400": {
                        "description": "Ошибка валидации данных запроса",
                        "schema": {
                            "$ref": "#/definitions/avito-rest-api_internal_error.ErrUserValidationError"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/avito-rest-api_internal_error.ErrInternalServerError"
                        }
                    }
                }
            }
        },
        "/api/v1/users/{id}": {
            "get": {
                "description": "Возвращает пользователя с указанным ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Получить пользователя по ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID пользователя",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Пользователь с указанным ID",
                        "schema": {
                            "$ref": "#/definitions/internal_controller_http_v1.GetUserByIDResponse"
                        }
                    },
                    "400": {
                        "description": "Ошибка валидации данных запроса",
                        "schema": {
                            "$ref": "#/definitions/avito-rest-api_internal_error.ErrUserValidationError"
                        }
                    },
                    "404": {
                        "description": "Пользователь с указанным ID не был найден",
                        "schema": {
                            "$ref": "#/definitions/avito-rest-api_internal_error.ErrUserNotFound"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/avito-rest-api_internal_error.ErrInternalServerError"
                        }
                    }
                }
            }
        },
        "/api/v1/users/{id}/withSegments": {
            "get": {
                "description": "Возвращает пользователя с указанным ID, включая в тело ответа список сегментов, в которые пользователь входит на момент совершения запроса",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Получить пользователя с его сегментами по ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID пользователя",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Пользователь с его активными сегментами",
                        "schema": {
                            "$ref": "#/definitions/internal_controller_http_v1.GetUserByIDWithSegmentsResponse"
                        }
                    },
                    "400": {
                        "description": "Ошибка валидации данных запроса",
                        "schema": {
                            "$ref": "#/definitions/avito-rest-api_internal_error.ErrUserValidationError"
                        }
                    },
                    "404": {
                        "description": "Пользователь с указанным ID не был найден",
                        "schema": {
                            "$ref": "#/definitions/avito-rest-api_internal_error.ErrUserNotFound"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/avito-rest-api_internal_error.ErrInternalServerError"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "avito-rest-api_internal_entity.User": {
            "type": "object",
            "properties": {
                "age": {
                    "type": "integer",
                    "example": 26
                },
                "is_deleted": {
                    "type": "boolean",
                    "example": false
                },
                "lastname": {
                    "type": "string",
                    "example": "Иванов"
                },
                "name": {
                    "type": "string",
                    "example": "Михаил"
                },
                "sex": {
                    "description": "Пол, 0 - мужской, 1 - женский",
                    "type": "integer",
                    "enum": [
                        0,
                        1
                    ],
                    "example": 0
                },
                "sex_text": {
                    "type": "string",
                    "enum": [
                        "мужской",
                        "женский"
                    ],
                    "example": "мужской"
                },
                "user_id": {
                    "type": "integer",
                    "example": 16
                }
            }
        },
        "avito-rest-api_internal_entity.UserSegmentInformation": {
            "type": "object",
            "properties": {
                "end_date": {
                    "description": "Дата выхода пользователя из сегмента",
                    "type": "string",
                    "example": ""
                },
                "information_id": {
                    "description": "ID записи в ассоциативной таблице, связывающей пользователей с сегментами",
                    "type": "integer",
                    "example": 179
                },
                "name": {
                    "description": "Наименование сегмента",
                    "type": "string",
                    "example": "AVITO_MUSIC_SERVICE"
                },
                "segment_id": {
                    "description": "ID сегмента",
                    "type": "integer",
                    "example": 43
                },
                "start_date": {
                    "description": "Дата добавления пользователя в сегмент",
                    "type": "string",
                    "example": "15:27:32 01.09.2023"
                },
                "user_id": {
                    "description": "ID пользователя, который входит в сегмент",
                    "type": "integer",
                    "example": 16
                }
            }
        },
        "avito-rest-api_internal_entity.UserWithSegments": {
            "type": "object",
            "properties": {
                "segments": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/avito-rest-api_internal_entity.UserSegmentInformation"
                    }
                },
                "user": {
                    "$ref": "#/definitions/avito-rest-api_internal_entity.User"
                }
            }
        },
        "avito-rest-api_internal_error.ErrInternalServerError": {
            "type": "object",
            "properties": {
                "comment": {
                    "type": "string"
                },
                "location": {
                    "type": "string"
                },
                "originErrorText": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                }
            }
        },
...
        "avito-rest-api_internal_error.ErrUserNotFound": {
            "type": "object",
            "properties": {
                "comment": {
                    "type": "string"
                },
                "location": {
                    "type": "string"
                },
                "originErrorText": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                }
            }
        },
        "avito-rest-api_internal_error.ErrUserValidationError": {
            "type": "object",
            "properties": {
                "comment": {
                    "type": "string"
                },
                "location": {
                    "type": "string"
                },
                "originErrorText": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                }
            }
        },
...
        "avito-rest-api_internal_service.UserCreateInput": {
            "type": "object",
            "required": [
                "age",
                "lastname",
                "name",
                "sex"
            ],
            "properties": {
                "age": {
                    "type": "integer",
                    "example": 26
                },
                "lastname": {
                    "type": "string",
                    "example": "Иванов"
                },
                "name": {
                    "type": "string",
                    "example": "Михаил"
                },
                "sex": {
                    "description": "Пол, 0 - мужской, 1 - женский",
                    "type": "integer",
                    "enum": [
                        0,
                        1
                    ],
                    "example": 0
                }
            }
        },
        "internal_controller_http_v1.AddUserToSegmentsInput": {
            "type": "object",
            "required": [
                "id"
            ],
            "properties": {
                "id": {
                    "description": "Идентификатор пользователя",
                    "type": "integer",
                    "example": 16
                },
                "segments": {
                    "description": "Сегменты, в которые необходимо добавить пользователя",
                    "type": "array",
                    "items": {
                        "type": "object",
                        "required": [
                            "name"
                        ],
                        "properties": {
                            "end_date": {
                                "description": "Необязательное поле, если отсутствует, значит дата выхода пользователя из сегмента не определена",
                                "type": "string",
                                "example": "10:00:00 25.09.2023"
                            },
                            "name": {
                                "description": "Наименование сегмента, в который необходимо добавить пользователя",
                                "type": "string",
                                "example": "AVITO_MUSIC_SERVICE"
                            }
                        }
                    }
                }
            }
        },
        "internal_controller_http_v1.AddUserToSegmentsResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "user 16 was successfully added to the segments"
                }
            }
        },
...
        "internal_controller_http_v1.DeleteUserFromSegmentsInput": {
            "type": "object",
            "required": [
                "id"
            ],
            "properties": {
                "id": {
                    "description": "Идентификатор пользователя",
                    "type": "integer",
                    "example": 16
                },
                "segments": {
                    "description": "Сегменты, из которых необходимо удалить пользователя",
                    "type": "array",
                    "items": {
                        "type": "object",
                        "required": [
                            "name"
                        ],
                        "properties": {
                            "name": {
                                "description": "Наименование сегмента",
                                "type": "string",
                                "example": "AVITO_MUSIC_SERVICE"
                            }
                        }
                    }
                }
            }
        },
        "internal_controller_http_v1.DeleteUserFromSegmentsResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "user 179 was successfully removed from segments"
                }
            }
        },
...
        "internal_controller_http_v1.GetAllUsersResponse": {
            "type": "object",
            "properties": {
                "users": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/avito-rest-api_internal_entity.User"
                    }
                }
            }
        },
        "internal_controller_http_v1.GetAllUsersWithSegmentsResponse": {
            "type": "object",
            "properties": {
                "users": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/avito-rest-api_internal_entity.UserWithSegments"
                    }
                }
            }
        },
...
        "internal_controller_http_v1.GetUserByIDResponse": {
            "type": "object",
            "properties": {
                "user": {
                    "$ref": "#/definitions/avito-rest-api_internal_entity.User"
                }
            }
        },
        "internal_controller_http_v1.GetUserByIDWithSegmentsResponse": {
            "type": "object",
            "properties": {
                "user": {
                    "$ref": "#/definitions/avito-rest-api_internal_entity.UserWithSegments"
                }
            }
        },
...
        "internal_controller_http_v1.UserCreateResponse": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer",
                    "example": 26
                }
            }
        }
    }
}